Fix NullPointerException thrown by ConsulAdvertiser
Initialize the healthCheckPath field to an empty Optional in
ConsulFactory. Otherwise, the ConsulAdvertise constructor
throws NullPointerException (at line 145) because the call to
configuration.getHealthCheckPath() returns a null instead of
an empty Optional.

Add test for this specific situation. Because all the existing
tests set the healthCheckPath, this test has to create a new
ConsulFactory and specifically NOT set the healthCheckPath.